modules/_requirements.txt
pytest
requests>=2.0
protobuf>=3.0.0
zeroconf>=0.21.0
casttube>=0.2.0
pychromecast
pybluez
bluepy  
scapy
access_points
python-nmap
pycairo
PyGObject
dbus-python
gobject
pycups
opencv-python
nfcpy
pyzbar
qrcode[pil]
readchar
pyscard
evdev
python-uinput
bs4
fleep
Pillow
pycrypto
netifaces
prettytable
libarchive-c
ctypescrypto
twisted
pcapy
dnspython
python-uinputce
paho-mqtt
shodan